Move mismatch fix method to task object itself.
So it can be reused, also fixed typo.
